Update README.md
messed up on the env and used my Twitter bot env. it is switched to the the right env now. if anyone used the old one to set up their env you can just erase the my_twitter_bot folder. I swapped the virtual environment to myvenv to have it stand out a bit better and make it less confusing.